requirements.txt
debugpy==1.8.20
fake-bpy-module-4.5==20260128
